contact-book/backend/requirements.txt
boto3
PyJWT
sqlite3
